requirements.txt
python-docx
selenium
openpyxl
pandas
undetected_chromedriver
setuptools
python3-distutils